requirements.txt
requests
pandas
colorama
python-dotenv

# google-cloud-aiplatform
# google-cloud-bigquery
# google-cloud-storage
# db-dtypes
# pandas-gbq

# ML Libraries for training pipeline
scikit-learn
xgboost
joblib
matplotlib
seaborn


feast
pyarrow